trainer/requirements.txt
torch==2.5.1+cu124 --extra-index-url https://download.pytorch.org/whl/cu124
torchvision==0.20.1+cu124 --extra-index-url https://download.pytorch.org/whl/cu124
torchaudio==2.5.1+cu124 --extra-index-url https://download.pytorch.org/whl/cu124
gsplat==1.4.0
numpy<2.0.0
Pillow
torchmetrics
tqdm
matplotlib
open3d
plyfile
ninja
setuptools
sympy>=1.13.0
